+++ src/hooks/Graph/useGraphFunc.tsx
@@ -4,7 +4,7 @@ import {
   SearchPageRankResult,
   SingularElementReturnValue,
 } from 'cytoscape';
-import * as graph from '@hooks/Graph/graphConstants';
+import * as graph from '@/util/Constants/graphConstants';
 
 const dimColor = '#dfe4ea';
 const DimStyle = {
